+++ packages/vtable/examples/menu.ts
@@ -146,6 +146,10 @@ export const menus = [
       {
         path: 'list',
         name: 'list-group-manyChild'
+      },
+      {
+        path: 'list',
+        name: 'list-tree-checkbox'
       }
     ]
   },
@@ -192,6 +196,10 @@ export const menus = [
       {
         path: 'pivot',
         name: 'pivot-tree-lazy-load'
+      },
+      {
+        path: 'pivot',
+        name: 'pivot-tree-hide'
       },
       {
         path: 'pivot',
@@ -520,6 +528,10 @@ export const menus = [
       {
         path: 'pivot-analysis',
         name: 'pivot-analysis-tree'
+      },
+      {
+        path: 'pivot-analysis',
+        name: 'pivot-analysis-hide'
       },
       {
         path: 'pivot-analysis',