+++ packages/vtable/src/ts-types/list-table/layout-map/api.ts
@@ -83,7 +83,7 @@ export interface HeaderData extends WidthData {
   hierarchyState?: HierarchyState;
   columnWidthComputeMode?: 'normal' | 'only-header' | 'only-body';
 
-  showSort?: boolean;
+  showSort?: boolean | ((args: { row: number; col: number; table: BaseTableAPI }) => boolean);
   sort?: SortOption;
 
   /**
@@ -181,7 +181,7 @@ export interface SeriesNumberColumnData extends WidthData {
   title?: string | (() => string);
   field?: FieldDef;
   // fieldKey?: FieldKeyDef;
-  format?: (col?: number, row?: number, table?: BaseTableAPI) => any;
+  format?: (col?: number, row?: number, table?: BaseTableAPI, originValue?: string | number) => any;
   // icon?: ColumnIconOption | ColumnIconOption[];
   icon?:
     | string